Ensure that the acl resource restricts the action attribute to valid values and refactor StringInSlice validation func
